python: add argparse.Action subclasses to flux.util
Problem: There are some common argparse.Action subclasses copied
in several flux utilities.

Reduce code duplication by moving these classes into flux.util.
